pyproject.toml
[project]
name = "yt-music-discovery"
version = "0.1.0"
description = ""
authors = [
    {name = "Jaime Rodriguez",email = "hi.i.am.jaime@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client (>=2.185.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "flask-limiter (>=4.0.0,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "yt_music_discovery", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
